+++ src/main.rs
@@ -17,26 +17,37 @@ fn main() {
 
             If you want to reorganize all classes in place, and change the files run with the `--write` flag
               rustywind --write .
+
+            If you want to run it on your STDIN, you can do:
+              echo \"<FILE CONTENTS>\" | rustywind --stdin
                          
             rustywind [FLAGS] <PATH>"))
         .arg(
             Arg::with_name("file_or_dir")
                 .value_name("PATH")
                 .help("A file or directory to run on")
+                .conflicts_with("stdin")
                 .index(1)
-                .required(true)
+                .required_unless("stdin")
                 .takes_value(true),
+        )
+        .arg(
+            Arg::with_name("stdin")
+                .long("stdin")
+                .conflicts_with_all(&["write", "file_or_dir", "dry-run"])
+                .required_unless("file_or_dir")
+                .help("Uses stdin instead of a file or folder")
         )
         .arg(
             Arg::with_name("write")
                 .long("write")
-                .conflicts_with("dry-run")
+                .conflicts_with_all(&["stdin", "dry-run"])
                 .help("Changes the files in place with the reorganized classes"),
         )
         .arg(
             Arg::with_name("dry_run")
                 .long("dry-run")
-                .conflicts_with("write")
+                .conflicts_with_all(&["stdin", "write"])
                 .help("Prints out the new file content with the sorted classes to the terminal"),
         )
         .arg(
@@ -49,6 +60,7 @@ fn main() {
     let options = Options::new_from_matches(&matches);
 
     match &options.write_mode {
+        WriteMode::ToStdOut => (),
         WriteMode::DryRun => println!(
             "\ndry run mode activated: here is a list of files that \
              would be changed when you run with the --write flag"
@@ -63,10 +75,24 @@ fn main() {
         ),
     }
 
-    options
-        .search_paths
-        .par_iter()
-        .for_each(|file_path| run_on_file_paths(&file_path, &options))
+    match &options.write_mode {
+        WriteMode::ToStdOut => {
+            let contents = options.stdin.clone().unwrap_or_else(|| "".to_string());
+
+            if rustywind::has_classes(&contents) {
+                let sorted_content = rustywind::sort_file_contents(contents, &options);
+                println!("{}", sorted_content)
+            } else {
+                println!("{}", contents);
+                eprintln!("[ERROR] No classes were found in STDIN");
+                std::process::exit(2)
+            }
+        }
+        _ => options
+            .search_paths
+            .par_iter()
+            .for_each(|file_path| run_on_file_paths(&file_path, &options)),
+    }
 }
 
 fn run_on_file_paths(file_path: &Path, options: &Options) {
@@ -76,6 +102,7 @@ fn run_on_file_paths(file_path: &Path, options: &Options) {
                 let sorted_content = rustywind::sort_file_contents(contents, options);
 
                 match &options.write_mode {
+                    WriteMode::ToStdOut => (),
                     WriteMode::DryRun => print_file_name(file_path, options),
                     WriteMode::ToFile => write_to_file(file_path, &sorted_content, options),
                     WriteMode::ToConsole => print_file_contents(&sorted_content),