+++ src/options.rs
@@ -1,5 +1,6 @@
 use clap::ArgMatches;
 use ignore::WalkBuilder;
+use std::io::Read;
 use std::path::{Path, PathBuf};
 
 #[derive(Debug)]
@@ -7,6 +8,7 @@ pub enum WriteMode {
     ToFile,
     DryRun,
     ToConsole,
+    ToStdOut,
 }
 
 #[derive(Debug)]
@@ -21,6 +23,7 @@ pub enum Sorter {
 
 #[derive(Debug)]
 pub struct Options {
+    pub stdin: Option<String>,
     pub write_mode: WriteMode,
     pub regex: FinderRegex,
     pub sorter: Sorter,
@@ -31,15 +34,35 @@ pub struct Options {
 
 impl Options {
     pub fn new_from_matches(matches: &ArgMatches) -> Options {
-        let starting_path = get_starting_path_from_matches(matches);
+        match matches.is_present("stdin") {
+            true => {
+                let mut buffer = String::new();
+                let mut stdin = std::io::stdin(); // We get `Stdin` here.
+                stdin.read_to_string(&mut buffer).unwrap();
 
-        Options {
-            write_mode: get_write_mode_from_matches(matches),
-            regex: FinderRegex::DefaultRegex,
-            sorter: Sorter::DefaultSorter,
-            starting_path: starting_path.to_owned(),
-            allow_duplicates: matches.is_present("allow-duplicates"),
-            search_paths: get_search_paths_from_starting_path(&starting_path),
+                Options {
+                    stdin: Some(buffer),
+                    write_mode: WriteMode::ToStdOut,
+                    regex: FinderRegex::DefaultRegex,
+                    sorter: Sorter::DefaultSorter,
+                    starting_path: PathBuf::new(),
+                    allow_duplicates: matches.is_present("allow-duplicates"),
+                    search_paths: vec![],
+                }
+            }
+            false => {
+                let starting_path = get_starting_path_from_matches(matches);
+
+                Options {
+                    stdin: None,
+                    write_mode: get_write_mode_from_matches(matches),
+                    regex: FinderRegex::DefaultRegex,
+                    sorter: Sorter::DefaultSorter,
+                    starting_path: starting_path.to_owned(),
+                    allow_duplicates: matches.is_present("allow-duplicates"),
+                    search_paths: get_search_paths_from_starting_path(&starting_path),
+                }
+            }
         }
     }
 }